qwen-tts/requirements.txt
dashscope>=1.14.0
requests>=2.25.0
pygame>=2.0.0
python-dotenv>=0.19.0
pathlib2>=2.3.0; python_version < "3.4" 